Add checking ability and fix matrix list bug
The code now can check if an inutted matrix is correct, but the
checking function seems to have a few bugs and occassionally says
correct when it should be incorrect. The bug that was fixed was a
bug where, if you pressed too many times on the check button
(and in turn ran the ref or rref function too many times), the
convert code would crash.